221011 feat/chat directory change
folder refactoring
