--- src/main/java/com/innovation/backend/chat/storage/MessageStorage.java
@@ -1,4 +1,0 @@
-package com.innovation.backend.chat.storage;
-
-public class MessageStorage {
-}
