+++ src/main/java/com/innovation/backend/domain/Chat/dto/response/MessageResponseDto.java
@@ -1,4 +1,4 @@
-package com.innovation.backend.chat.dto.response;
+package com.innovation.backend.domain.Chat.dto.response;
 
 import lombok.Getter;
 import lombok.NoArgsConstructor;
