+++ src/main/java/com/innovation/backend/domain/Chat/repository/MessageRepository.java
@@ -1,6 +1,6 @@
-package com.innovation.backend.chat.repository;
+package com.innovation.backend.domain.Chat.repository;
 
-import com.innovation.backend.chat.entity.Message;
+import com.innovation.backend.domain.Chat.entity.Message;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface MessageRepository extends JpaRepository<Message, String> {
