+++ src/main/java/com/innovation/backend/domain/Chat/service/ChatRoomService.java
@@ -1,9 +1,9 @@
-package com.innovation.backend.chat.service;
+package com.innovation.backend.domain.Chat.service;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
 @Service
 @RequiredArgsConstructor
-public class MessageService {
+public class ChatRoomService {
 }
