+++ src/main/java/com/innovation/backend/domain/Chat/storage/MessageStorage.java
@@ -1,0 +1,4 @@
+package com.innovation.backend.domain.Chat.storage;
+
+public class MessageStorage {
+}
